finalizing the ConsumptionServiceImplementation logic

+++ src/Services/ConsumptionService.java
@@ -9,6 +9,6 @@ import java.time.LocalDate;
 public interface ConsumptionService {
     boolean add(ConsumptionType consumptionType , int id , User user , float carbonQuantity , int intParam ,  AllTypesOfConsumption consumptionImpactType , LocalDate startDate , LocalDate endDate);
 
-    float calculateAverageOfConsumptionWithinARange(User user, LocalDate startDate, LocalDate endDate);
+    double calculateAverageOfConsumptionWithinARange(User user, LocalDate startDate, LocalDate endDate);
 
 }
